% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{addRegistrant}
\alias{addRegistrant}
\title{Register Participants for a Meeting}
\usage{
addRegistrant(
  meeting_id,
  email,
  first_name,
  token_cache = "~/.zoom_token_cache"
)
}
\arguments{
\item{meeting_id}{ID of the meeting to register the participant for.}

\item{email}{Registrant email.}

\item{first_name}{Registrant first name.}

\item{token_cache}{Path to token cache}
}
\description{
Post to the /meetings/{meeting_id}/registrants endpoint.
}
\examples{
addRegistrant()
}
\keyword{registration}
