% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{createMeeting}
\alias{createMeeting}
\title{Create a new meeting.}
\usage{
createMeeting(
  topic,
  start_time,
  duration = 60,
  type = 2,
  pre_schedule = FALSE,
  timezone = "America/New_York",
  settings = list(),
  template_id = "",
  user_id = "me",
  token_cache = "~/.zoom_token_cache"
)
}
\arguments{
\item{topic}{Topic/title of meeting.}

\item{start_time}{Starting time of meeting.  For example, "2021-08-19T17:10:00".}

\item{duration}{Duration of meeting in minutes.  Defaults to 60.}

\item{type}{Type of meeting.  Defaults to 2.}

\item{pre_schedule}{Pre-schedule meeting? Defaults to FALSE.}

\item{timezone}{Time zone to schedule meeting.  Defaults to "America/New_York".}

\item{settings}{An optional named list of settings.}

\item{template_id}{Template ID for creating meeting.  Defaults to "".}

\item{user_id}{Specify user ID to create a meeting for.  Defaults to "me".}

\item{token_cache}{Path to token cache}
}
\description{
Post to the /users/{user_id}/meetings endpoint.
}
\examples{
createMeeting()
}
\keyword{meetings}
