% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getMeetingPollsResults}
\alias{getMeetingPollsResults}
\title{Get Meeting Poll Results}
\usage{
getMeetingPollsResults(meeting_id, token = NULL)
}
\arguments{
\item{meeting_id}{Specify meeting ID whose recordings to lookup.}
}
\description{
Query the /past_meetings/{meeting_id}/polls endpoint.
}
\examples{
getMeetingPollsResults()
}
\keyword{polls}
