% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getMyMeetings}
\alias{getMyMeetings}
\title{Get User Meetings}
\usage{
getMyMeetings(user_id = "me", token = NULL)
}
\arguments{
\item{user_id}{Specify ID of user whose meetings to lookup.  Defaults to "me".}
}
\description{
Query the /users/{user_id}/meetings endpoint.
}
\examples{
getMyMeetings()
}
\keyword{users}
