% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{listMeetingPolls}
\alias{listMeetingPolls}
\title{List Meeting Polls}
\usage{
listMeetingPolls(meeting_id, token = NULL)
}
\arguments{
\item{meeting_id}{Specify meeting ID whose polls to lookup.}
}
\description{
Query the /users/{meeting_id}/polls endpoint.
}
\examples{
listMeetingPolls()
}
\keyword{polls}
