% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{listRecordings}
\alias{listRecordings}
\title{List Cloud Recordings}
\usage{
listRecordings(user_id = "me", token = NULL)
}
\arguments{
\item{meeting_id}{Specify user ID whose recordings to lookup.}
}
\description{
Query the /users/{user_id}/recordings endpoint.
}
\examples{
listRecordings()
}
\keyword{recordings}
