% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{listUser}
\alias{listUser}
\title{Get User}
\usage{
listUser(user_id = "me", token_cache = "~/.zoom_token_cache")
}
\arguments{
\item{user_id}{Specify ID of user to lookup.  Defaults to "me".}

\item{token_cache}{Path to token cache}
}
\description{
Query the /users/ endpoint.
}
\examples{
user()
}
\keyword{users}
