% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{loadToken}
\alias{loadToken}
\title{Load Saved Zoom Token}
\usage{
loadToken()
}
\description{
This function allows you to load a Zoom API token previously saved in the home directory.
}
\examples{
loadToken()
}
\keyword{Auth}
