% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{makeToken}
\alias{makeToken}
\title{Make a Zoom Token}
\usage{
makeToken(client_id, client_secret, redirect_uri, app_name = "", save = TRUE)
}
\arguments{
\item{client_id}{Client ID associated with your Zoom app.}

\item{client_secret}{Client secret associated with your Zoom app.}

\item{redirect_uri}{Redirect URI associated with your Zoom app.}

\item{app_name}{Optional app name, for organizational purposes.}

\item{save}{Save the token in the home directory?  Defaults to TRUE.}
}
\description{
This function allows you to make a Zoom API token from your app.
}
\examples{
makeToken()
}
\keyword{Auth}
