% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{meetingInfo}
\alias{meetingInfo}
\title{Get Meeting Info}
\usage{
meetingInfo(meeting_id, token_cache = "~/.zoom_token_cache")
}
\arguments{
\item{meeting_id}{Specify ID of meeting to lookup.}

\item{token_cache}{Path to token cache}
}
\description{
Query the /meetings/{meeting_id} endpoint.
}
\examples{
meetingInfo()
}
\keyword{meetings}
