% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{listMeetingPolls}
\alias{listMeetingPolls}
\title{List Meeting Polls}
\usage{
listMeetingPolls(meeting_id, token_cache = "~/.zoom_token_cache")
}
\arguments{
\item{meeting_id}{Specify meeting ID whose polls to lookup.}

\item{token_cache}{Path to token cache}
}
\description{
Query the /users/{meeting_id}/polls endpoint.
}
\examples{
listMeetingPolls()
}
\keyword{polls}
