% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{listMeetingRegistrants}
\alias{listMeetingRegistrants}
\title{List Meeting Registrants}
\usage{
listMeetingRegistrants(meeting_id, token = NULL)
}
\arguments{
\item{meeting_id}{Specify user ID whose recordings to lookup.  Defaults to "me".}
}
\description{
Query the /meetings/{meeting_id}/registrants endpoint.
}
\examples{
listMeetingRegistrants()
}
\keyword{registration}
