% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{listRecordings}
\alias{listRecordings}
\title{List Cloud Recordings}
\usage{
listRecordings(user_id = "me", token_cache = "~/.zoom_token_cache")
}
\arguments{
\item{token_cache}{Path to token cache}

\item{meeting_id}{Specify user ID whose recordings to lookup.}
}
\description{
Query the /users/{user_id}/recordings endpoint.
}
\examples{
listRecordings()
}
\keyword{recordings}
