% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{listUser}
\alias{listUser}
\title{Get User}
\usage{
listUser(user_id = "me", token = NULL)
}
\arguments{
\item{user_id}{Specify ID of user to lookup.  Defaults to "me".}
}
\description{
Query the /users/ endpoint.
}
\examples{
user()
}
\keyword{users}
