% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{meetingInfo}
\alias{meetingInfo}
\title{Get Meeting Info}
\usage{
meetingInfo(meeting_id, token = NULL)
}
\arguments{
\item{meeting_id}{Specify ID of meeting to lookup.}
}
\description{
Query the /meetings/{meeting_id} endpoint.
}
\examples{
meetingInfo()
}
\keyword{meetings}
